backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Request handling
python-multipart>=0.0.6
aiofiles>=23.2.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
greenlet>=3.0.0

# Settings management
pydantic-settings>=2.0.0

# AI/LLM providers
google-genai>=0.3.0
google-api-core>=2.0.0
langchain-text-splitter>=0.0.1
openai

# Document parsing
python-pptx>=0.6.21
pdf2image>=1.17.0
pdfplumber
pytesseract>=0.3.10

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.26.0
